requirements.txt
aiocache==0.12.0
asyncpg==0.27.0
DateTime==4.3
dadata==21.10.1
fastapi==0.95.0
pydantic==1.8.2
pytest==6.2.5
python-dotenv==0.19.0
uvicorn==0.15.0


